% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{eigenMSNorm}
\alias{eigenMSNorm}
\title{EigenMS Normalization}
\usage{
eigenMSNorm(se, ain = "log2", aout = "EigenMS")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default log2)}

\item{aout}{String which assay should be used to save normalized data (default eigenMS)}
}
\value{
SummarizedExperiment containing the EigenMS normalized data as assay
}
\description{
EigenMS Normalization
}
