% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerDE.R
\name{getIndexList}
\alias{getIndexList}
\title{Return list containing vector positions of values in string}
\usage{
getIndexList(targetVector)
}
\arguments{
\item{targetVector}{vector of sample groups}
}
\value{
indexList List where key is condition level and values are indices
  for the condition
}
\description{
Return list containing vector positions of values in string
}
\keyword{internal}
