% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{globalIntNorm}
\alias{globalIntNorm}
\title{Total Intensity Normalization
Intensities of each variable in a sample are divided with the sum of intensities
of all variables in the sample and multiplied with the median or mean of sum of intensities
of all variables in all samples. Raw data is taken as input (not log2).}
\usage{
globalIntNorm(se, ain = "raw", aout = "GlobalMedian", type = "median")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{ain}{String which assay should be used as input (default raw)}

\item{aout}{String which assay should be used to save normalized data (default TotalInt_mean or TotalInt_median)}

\item{type}{String whether to use median or mean to calculate the scaling factor}
}
\value{
SummarizedExperiment containing the total intensity normalized data as assay
}
\description{
Total Intensity Normalization
Intensities of each variable in a sample are divided with the sum of intensities
of all variables in the sample and multiplied with the median or mean of sum of intensities
of all variables in all samples. Raw data is taken as input (not log2).
}
