% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{globalMeanNorm}
\alias{globalMeanNorm}
\title{Total Intensity Normalization Using Mean For Calculation of Scaling Factors}
\usage{
globalMeanNorm(se, ain = "raw", aout = "GlobalMean")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{ain}{String which assay should be used as input (default raw)}

\item{aout}{String which assay should be used to save normalized data (default GlobalMean)}
}
\value{
SummarizedExperiment containing the total intensity normalized data as assay
}
\description{
Total Intensity Normalization Using Mean For Calculation of Scaling Factors
}
