% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{globalMedianNorm}
\alias{globalMedianNorm}
\title{Total Intensity Normalization Using Median For Calculation of Scaling Factors}
\usage{
globalMedianNorm(se, ain = "raw", aout = "GlobalMedian")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default raw)}

\item{aout}{String which assay should be used to save normalized data (default GlobalMedian)}
}
\value{
SummarizedExperiment containing the total intensity normalized data as assay
}
\description{
Total Intensity Normalization Using Median For Calculation of Scaling Factors
}
