% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{impute_se}
\alias{impute_se}
\title{Method to impute SummarizedExperiment}
\usage{
impute_se(se, ain = NULL, condition = NULL)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics dataset}

\item{ain}{Vector of strings which assay should be used as input (default NULL).
If NULL then all normalization of the se object are plotted next to each other.}

\item{condition}{name of column of colData(se) representing the conditions of the data}
}
\value{
SummarizedExperiment with imputed intensities
}
\description{
Method to impute SummarizedExperiment
}
