% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{irsNorm}
\alias{irsNorm}
\title{Internal Reference Scaling Normalization
IRS makes different measurements of the same thing all exactly the same and puts
all of the intensities on the same scale.}
\usage{
irsNorm(se, ain = "raw", aout = "IRS")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default raw)}

\item{aout}{String which assay should be used to save normalized data (default IRS)}
}
\value{
SummarizedExperiment containing the IRS normalized data as assay
}
\description{
Internal Reference Scaling Normalization
IRS makes different measurements of the same thing all exactly the same and puts
all of the intensities on the same scale.
}
