% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{limmaNorm}
\alias{limmaNorm}
\title{limma::removeBatchEffects (limBE)}
\usage{
limmaNorm(se, ain = "log2", aout = "limBE")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default log2)}

\item{aout}{String which assay should be used to save normalized data (default limBE)}
}
\value{
SummarizedExperiment containing the limBE normalized data as assay
}
\description{
limma::removeBatchEffects (limBE)
}
