% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{loessCycNorm}
\alias{loessCycNorm}
\title{Cyclic Loess Normalization of limma (two samples of the data are MA transformed and normalized at a time, and all pairs of samples are iterated through)}
\usage{
loessCycNorm(se, ain = "log2", aout = "LoessCyc")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default log2)}

\item{aout}{String which assay should be used to save normalized data (default cyclicLoess)}
}
\value{
SummarizedExperiment containing the loessCyc normalized data as assay
}
\description{
Cyclic Loess Normalization of limma (two samples of the data are MA transformed and normalized at a time, and all pairs of samples are iterated through)
}
