% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{loessFNorm}
\alias{loessFNorm}
\title{Fast Loess Normalization of limma (using mean intensities over all the samples as its reference A sample)}
\usage{
loessFNorm(se, ain = "log2", aout = "LoessF")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default log2)}

\item{aout}{String which assay should be used to save normalized data (default loessF)}
}
\value{
SummarizedExperiment containing the loessF normalized data as assay
}
\description{
Fast Loess Normalization of limma (using mean intensities over all the samples as its reference A sample)
}
