% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{medianNorm}
\alias{medianNorm}
\title{Median Normalization
The intensity of each protein group in a given sample is divided by the median of the
intensities of all protein groups in that sample and then multiplied by the mean of
median of sum of intensities of all protein groups in all samples.}
\usage{
medianNorm(se, ain = "log2", aout = "Median")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default log2)}

\item{aout}{String which assay should be used to save normalized data (default mean)}
}
\value{
SummarizedExperiment containing the median normalized data as assay
}
\description{
Median Normalization
The intensity of each protein group in a given sample is divided by the median of the
intensities of all protein groups in that sample and then multiplied by the mean of
median of sum of intensities of all protein groups in all samples.
}
