% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{normalize_se}
\alias{normalize_se}
\title{Normalize SummarizedExperiment object using single normalization methods or specified combinations of normalization methods}
\usage{
normalize_se(se, methods, combination_pattern = "_on_")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{methods}{Vector of normalization methods to apply for normalizing the proteomics data of the SummarizedExperiment object (identifier of normalization methods can be retrieved using get_all_normalization_methods())}

\item{combination_pattern}{String specifying how normalization methods are combined. For instance, methods = c("IRS", "Median_on_IRS"), combination_pattern = "_on_".}
}
\value{
SummarizedExperiment object with normalized data saved as assays
}
\description{
Normalize SummarizedExperiment object using single normalization methods or specified combinations of normalization methods
}
