% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{normalize_se_single}
\alias{normalize_se_single}
\title{Normalize SummarizedExperiment object using different normalization methods}
\usage{
normalize_se_single(se, methods = NULL, gamma.0 = 0.5)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{methods}{Vector of normalization methods to apply for normalizing the proteomics data of the SummarizedExperiment object (identifier of normalization methods can be retrieved using get_all_normalization_methods())}

\item{gamma.0}{Numeric representing the exponent of the weighted density of RobNorm normalization. When the sample size is small, the fitted population of some proteins could be locally trapped such that the variance of those proteins was very small under a large gamma. To avoid this, a small gamma is recommended. When sample size smaller than 40, then set gamma to 0.5 or 0.1.}
}
\value{
SummarizedExperiment object with normalized data saved as assays
}
\description{
Normalize SummarizedExperiment object using different normalization methods
}
