% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysis.R
\name{perform_limma}
\alias{perform_limma}
\title{Fitting a linear model using limma}
\usage{
perform_limma(data, condition_vector, comparisons, covariate = NULL)
}
\arguments{
\item{data}{Data table of intensities (rows = proteins, cols = samples)}

\item{condition_vector}{Vector of experimental design specifying the condition(s) to compare}

\item{comparisons}{Vector of comparisons that are performed in the DE analysis (from specify_comparisons method)}

\item{covariate}{String specifying which column to include as covariate into limma}
}
\value{
eBayes object
}
\description{
Fitting a linear model using limma
}
