% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPlots.R
\name{plot_upset_DE}
\alias{plot_upset_DE}
\title{Upset plots of DE results of the different normalization methods}
\usage{
plot_upset_DE(de_res, ain = NULL, comparisons = NULL, min_degree = 2)
}
\arguments{
\item{de_res}{data table resulting of run_DE}

\item{ain}{Vector of strings of normalization methods to visualize (must be valid normalization methods saved in de_res)}

\item{comparisons}{Vector of comparisons (must be valid comparisons saved in de_res)}

\item{min_degree}{Minimal degree of an intersection for it to be included}
}
\value{
list of list of ComplexUpset objects and list of data tables with intersections (one for each comparison)
}
\description{
Upset plots of DE results of the different normalization methods
}
