% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizR.R
\name{prepare_data_for_harmonizR}
\alias{prepare_data_for_harmonizR}
\title{Helper function to prepare input data for the HarmonizR function}
\usage{
prepare_data_for_harmonizR(se, ain = "raw", batch_column = NULL)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{ain}{String which assay should be used as input (default raw)}

\item{batch_column}{column name of batch (if NULL, batch saved in SummarizedExperiment will be taken)}
}
\value{
list of two data frames
}
\description{
Helper function to prepare input data for the HarmonizR function
}
