% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{quantileNorm}
\alias{quantileNorm}
\title{Quantile Normalization of preprocessCore package.
Forces distributions of the samples to be the same on the basis of the quantiles of the samples by replacing
each protein of a sample with the mean of the corresponding quantile.}
\usage{
quantileNorm(se, ain = "log2", aout = "Quantile")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default log2)}

\item{aout}{String which assay should be used to save normalized data (default quantile)}
}
\value{
SummarizedExperiment containing the quantile normalized data as assay
}
\description{
Quantile Normalization of preprocessCore package.
Forces distributions of the samples to be the same on the basis of the quantiles of the samples by replacing
each protein of a sample with the mean of the corresponding quantile.
}
