% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{rlrMACycNorm}
\alias{rlrMACycNorm}
\title{Cyclic Linear Regression Normalization on MA Transformed Data (no reference, but MA transformation and normalization
of samples done pairwise between two samples, A = average of two samples, M = difference, process iterated through all samples pairs).}
\usage{
rlrMACycNorm(se, ain = "log2", aout = "RlrMACyc", iterations = 3)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default log2)}

\item{aout}{String which assay should be used to save normalized data (default rlrMACyc)}

\item{iterations}{Number of cyclic iterations to be performed (default 3)}
}
\value{
SummarizedExperiment containing the rlrMACyc normalized data as assay

SummarizedExperiment containing the rlrMACyc normalized data as assay
}
\description{
Cyclic Linear Regression Normalization on MA Transformed Data (no reference, but MA transformation and normalization
of samples done pairwise between two samples, A = average of two samples, M = difference, process iterated through all samples pairs).
}
