% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{rlrMANorm}
\alias{rlrMANorm}
\title{Linear Regression Normalization on MA Transformed Data (similar to Rlr, but data are MA transformed before normalization,
(A = median sample, M = difference of that sample to A)}
\usage{
rlrMANorm(se, ain = "log2", aout = "RlrMA")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default log2)}

\item{aout}{String which assay should be used to save normalized data (default rlrMA)}
}
\value{
SummarizedExperiment containing the rlrMA normalized data as assay

SummarizedExperiment containing the rlrMA normalized data as assay
}
\description{
Linear Regression Normalization on MA Transformed Data (similar to Rlr, but data are MA transformed before normalization,
(A = median sample, M = difference of that sample to A)
}
