% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{rlrNorm}
\alias{rlrNorm}
\title{Robust Linear Regression Normalization of NormalyzerDE (uses median values over all samples as reference
sample to which all the other samples in the data are normalized to)}
\usage{
rlrNorm(se, ain = "log2", aout = "Rlr")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default log2)}

\item{aout}{String which assay should be used to save normalized data (default rlr)}
}
\value{
SummarizedExperiment containing the rlr normalized data as assay
}
\description{
Robust Linear Regression Normalization of NormalyzerDE (uses median values over all samples as reference
sample to which all the other samples in the data are normalized to)
}
