% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{robNorm}
\alias{robNorm}
\title{RobNorm}
\usage{
robNorm(se, ain = "log2", aout = "RobNorm", gamma.0 = 0.1)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default log2)}

\item{aout}{String which assay should be used to save normalized data (default robNorm)}

\item{gamma.0}{Numeric representing the exponent of the weighted density. When the sample size
is small, the fitted population of some proteins could be locally trapped such
that the variance of those proteins was very small under a large gamma. To avoid
this, a small gamma is recommended. When sample size smaller than 40, then set
gamma to 0.5 or 0.1.}
}
\value{
SummarizedExperiment containing the robNorm normalized data as assay
}
\description{
RobNorm
}
