% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizR.R
\name{run_harmonizR}
\alias{run_harmonizR}
\title{HarmonizR proteomics data}
\usage{
run_harmonizR(
  se,
  ain = "raw",
  aout = "HarmonizR_ComBat",
  algorithm = "ComBat",
  combat_mode = 1
)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{ain}{String which assay should be used as input (default raw)}

\item{aout}{String which assay should be used to save normalized data (default HarmonizR_ComBat)}

\item{algorithm}{String specifying which algorithm should be used (ComBat or limma)}

\item{combat_mode}{Integer speciying which ComBat mode should be executed (1,2,3,4)}
}
\value{
SummarizedExperiment containing the HarmonizR data as assay
}
\description{
HarmonizR proteomics data
}
