% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se}
\alias{se}
\title{Example real-world proteomics data set}
\format{
## `se`
}
\source{
Biadglegne et al. Mycobacterium tuberculosis A↵ects Protein and Lipid Content of Circulating Exosomes in Infected Patients Depending on Tuberculosis Disease State. Biomedicines 10.4 (Mar. 2022), p. 783. doi: 10.3390/ biomedicines10040783.
}
\usage{
se
}
\description{
A \code{SummarizedExperiment} containing the raw and log2-scaled data of 301 proteins measured in 20 samples
}
\keyword{datasets}
