% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{tmmNorm}
\alias{tmmNorm}
\title{Weighted Trimmed Mean of M Values (TMM) Normalization}
\usage{
tmmNorm(se, ain = "raw", aout = "TMM")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default raw)}

\item{aout}{String which assay should be used to save normalized data (default TMM)}
}
\value{
SummarizedExperiment containing the TMM normalized data as assay
}
\description{
Weighted Trimmed Mean of M Values (TMM) Normalization
}
