% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{medianAbsDevNorm}
\alias{medianAbsDevNorm}
\title{Median Absolute Deviation Normalization (substracts the median and divides the data by the median absolute deviation (MAD))}
\usage{
medianAbsDevNorm(se, ain = "log2", aout = "MAD")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default log2)}

\item{aout}{String which assay should be used to save normalized data (default mad)}
}
\value{
SummarizedExperiment containing the MAD normalized data as assay
}
\description{
Median Absolute Deviation Normalization (substracts the median and divides the data by the median absolute deviation (MAD))
}
