% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizR.R
\name{run_harmonizR}
\alias{run_harmonizR}
\title{HarmonizR proteomics data}
\usage{
run_harmonizR(
  se,
  ain = "raw",
  aout = "HarmonizR_ComBat",
  batch_column = NULL,
  algorithm = "ComBat",
  combat_mode = 1
)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{ain}{String which assay should be used as input (default raw)}

\item{aout}{String which assay should be used to save normalized data (default HarmonizR_ComBat)}

\item{batch_column}{column name of batch (if NULL, batch saved in SummarizedExperiment will be taken)}

\item{algorithm}{String specifying which algorithm should be used (ComBat or limma)}

\item{combat_mode}{Integer speciying which ComBat mode should be executed (1,2,3,4)}
}
\value{
SummarizedExperiment containing the HarmonizR data as assay
}
\description{
HarmonizR proteomics data
}
\details{
combat_mode = 1 -> par.prior = TRUE, mean.only = FALSE
combat_mode = 2 -> par.prior = TRUE, mean.only = TRUE
combat_mode = 3 -> par.prior = FALSE, mean.only = FALSE
combat_mode = 4 -> par.prior = FALSE, mean.only = TRUE
par.prior -> TRUE indicates parametric adjustments will be used
mean.only -> TRUE indicates that ComBat only corrects the mean of the batch effect (no scale adjustment)
}
