% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{vsnNorm}
\alias{vsnNorm}
\title{Variance Stabilization Normalization of limma package}
\usage{
vsnNorm(se, ain = "raw", aout = "VSN")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input (default raw)}

\item{aout}{String which assay should be used to save normalized data (default vsn)}
}
\value{
SummarizedExperiment containing the vsn normalized data as assay
}
\description{
Variance Stabilization Normalization of limma package
}
